комментарии jsdoc
