--- src/mock/data.js
@@ -1,61 +1,0 @@
-/**
- * Хранение данных для генерации информации о фильмах
- * @exports titles
- * @exports descriptionParts
- * @exports posters
- * @exports genres
- */
-const titles = [
-  `Friends`,
-  `The Big Bang Theory`,
-  `Revolutionary Road`,
-  `The social network`,
-  `The green mile`,
-  `The Great Gatsby`,
-  `Star Wars`,
-  `Leon`,
-  `Game of Trones`,
-  `The Matrix`,
-  `Forrest Gamp`,
-  `A beautiful mind`,
-  `Harry Potter`,
-  `Frozen`,
-  `Lalaland`,
-];
-
-const descriptionParts = [
-  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
-  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
-  `Fusce tristique felis at fermentum pharetra.`,
-  `Aliquam id orci ut lectus varius viverra.`,
-  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
-  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
-  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
-  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
-  `Aliquam erat volutpat.`,
-  `Nunc fermentum tortor ac porta dapibus.`,
-  `In rutrum ac purus sit amet tempus.`,
-];
-
-const posters = [
-  `made-for-each-other.png`,
-  `popeye-meets-sinbad.png`,
-  `sagebrush-trail.jpg`,
-  `santa-claus-conquers-the-martians.jpg`,
-  `the-dance-of-life.jpg`,
-  `the-great-flamarion.jpg`,
-  `the-man-with-the-golden-arm.jpg`,
-];
-
-const genres = new Set([
-  `comedy`,
-  `drama`,
-  `cartoon`,
-  `horror`,
-  `TVshow`,
-  `mystery`,
-  `film-noir`,
-  `biografy`
-]);
-
-export {titles, descriptionParts, posters, genres};
