+++ src/mock/popup-data.js
@@ -2,7 +2,7 @@
  * Модуль для генерации данных для попапа(детальная информация о фильме)
  * @exports getDetailInfo
  */
-import {titles, genres, descriptionParts, posters} from '../mock/data.js';
+// import {titles, genres, descriptionParts, posters} from '../mock/data.js';
 import {getRandomElement, getRandomPart, getRandomFloatNumber, getRandomNumber, getRandomBoolean} from '../util.js';
 
 const DESCRIPTION_MIN = 1;
@@ -22,6 +22,59 @@ const MIN_IN_HOUR = 60;
 const EARLIEST_COMMENTING_DATE = Date.parse(new Date(`2016-01-01`));
 const LATEST_COMMENTING_DATE = Date.now();
 const MAX_COMMENT_QUANTITTY = 5;
+
+const titles = [
+  `Friends`,
+  `The Big Bang Theory`,
+  `Revolutionary Road`,
+  `The social network`,
+  `The green mile`,
+  `The Great Gatsby`,
+  `Star Wars`,
+  `Leon`,
+  `Game of Trones`,
+  `The Matrix`,
+  `Forrest Gamp`,
+  `A beautiful mind`,
+  `Harry Potter`,
+  `Frozen`,
+  `Lalaland`,
+];
+
+const descriptionParts = [
+  `Lorem ipsum dolor sit amet, consectetur adipiscing elit.`,
+  `Cras aliquet varius magna, non porta ligula feugiat eget.`,
+  `Fusce tristique felis at fermentum pharetra.`,
+  `Aliquam id orci ut lectus varius viverra.`,
+  `Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante.`,
+  `Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum.`,
+  `Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui.`,
+  `Sed sed nisi sed augue convallis suscipit in sed felis.`,
+  `Aliquam erat volutpat.`,
+  `Nunc fermentum tortor ac porta dapibus.`,
+  `In rutrum ac purus sit amet tempus.`,
+];
+
+const posters = [
+  `made-for-each-other.png`,
+  `popeye-meets-sinbad.png`,
+  `sagebrush-trail.jpg`,
+  `santa-claus-conquers-the-martians.jpg`,
+  `the-dance-of-life.jpg`,
+  `the-great-flamarion.jpg`,
+  `the-man-with-the-golden-arm.jpg`,
+];
+
+const genres = new Set([
+  `comedy`,
+  `drama`,
+  `cartoon`,
+  `horror`,
+  `TVshow`,
+  `mystery`,
+  `film-noir`,
+  `biografy`
+]);
 
 const directors = [
   `David Fincher`,
@@ -116,6 +169,11 @@ export const getDetailInfo = () => {
   };
 };
 
+/**
+ *Генерирует необходимое количество моковых данных для карточек фильмов и попапов
+ * @param {Number} count
+ * @return {Array} массив содержащий объкты с данными по фильмам
+ */
 export const generateDetailInfo = (count) => {
 
   return new Array(count)
